+++ compiler/c_bridge/c_bridge.h
@@ -82,10 +82,10 @@
 /// they are implemented in c_bridge.c
 
 // stdio.h
-C_BRIDGE_API int snprintf_wrap(char *restrict buffer, unsigned long long bufsz, const char *restrict format, ...);
+C_BRIDGE_API int snprintf_wrap(char *restrict buffer, size_t bufsz, const char *restrict format, ...);
 
 // string.h
-C_BRIDGE_API void *memset_wrap(void *dest, int ch, unsigned long long count);
+C_BRIDGE_API void *memset_wrap(void *dest, int ch, size_t count);
 
 // math.h
 C_BRIDGE_API double cos_wrap(double);
